refactor: move processCowAttrition to its own file
