+++ src/game-logic/reducers/index.test.js
@@ -1,7 +1,6 @@
 import { shapeOf, testCrop, testItem } from '../../test-utils'
 import {
   ACHIEVEMENT_COMPLETED,
-  COW_ATTRITION_MESSAGE,
   FERTILIZERS_PRODUCED,
   LOAN_INCREASED,
   LOAN_PAYOFF,
@@ -14,7 +13,6 @@ import {
   COW_HUG_BENEFIT,
   COW_MILK_RATE_SLOWEST,
   COW_FERTILIZER_PRODUCTION_RATE_SLOWEST,
-  COW_WEIGHT_MULTIPLIER_MINIMUM,
   MAX_ANIMAL_NAME_LENGTH,
   MAX_DAILY_COW_HUG_BENEFITS,
   MAX_LATEST_PEER_MESSAGES,
@@ -400,51 +398,6 @@ describe('computeStateForNextDay', () => {
     expect(firstRow[0].daysWatered).toBe(1)
     expect(firstRow[0].daysOld).toBe(1)
     expect(todaysNotifications).toBeEmpty()
-  })
-})
-
-describe('processCowAttrition', () => {
-  test('unfed cows leave', () => {
-    const unfedCow = generateCow({
-      name: 'unfed cow',
-      weightMultiplier: COW_WEIGHT_MULTIPLIER_MINIMUM,
-    })
-    const fedCow = generateCow({ name: 'fed cow', weightMultiplier: 1 })
-
-    const { cowInventory, newDayNotifications } = fn.processCowAttrition({
-      cowBreedingPen: { cowId1: null, cowId2: null, daysUntilBirth: -1 },
-      cowInventory: [unfedCow, fedCow],
-      inventory: [],
-      newDayNotifications: [],
-    })
-
-    expect(cowInventory).toEqual([fedCow])
-    expect(newDayNotifications).toEqual([
-      { message: COW_ATTRITION_MESSAGE`${unfedCow}`, severity: 'error' },
-    ])
-  })
-
-  test('used hugging machines are returned to inventory', () => {
-    const unfedCow = generateCow({
-      name: 'unfed cow',
-      weightMultiplier: COW_WEIGHT_MULTIPLIER_MINIMUM,
-    })
-    const unfedCowWithHuggingMachine = generateCow({
-      name: 'unfed cow',
-      weightMultiplier: COW_WEIGHT_MULTIPLIER_MINIMUM,
-      isUsingHuggingMachine: true,
-    })
-
-    const { cowInventory, inventory } = fn.processCowAttrition({
-      cowBreedingPen: { cowId1: null, cowId2: null, daysUntilBirth: -1 },
-      cowInventory: [unfedCow, unfedCowWithHuggingMachine],
-      inventory: [],
-      inventoryLimit: -1,
-      newDayNotifications: [],
-    })
-
-    expect(cowInventory).toEqual([])
-    expect(inventory).toEqual([{ id: huggingMachine.id, quantity: 1 }])
   })
 })
